Update how_to_use.md
I've identified a typo in the photo's dimensions near the bottom of your article.  You have: "`data-rokbox-size="210 213". Instead of 210 313
The corrected line would be displayed as: To set up a fixed-size RokBox2 popup, add `data-rokbox-size="210 313"` to the image modifiers where 210 is the width and 313 is the height of the popup image.

Hope that helps.
Dan..TheCardMan